Fix incorrect "this" use in calculator.model.sample.js
The issue involves Prep20_FinalAnnealedConcentration. Fortunately, because there is a currently a previous check for "that.nonStandard" (and the rest of that check is commented out) and the code then returns if that.nonStandard is true, it looks like the code can currently only reach the point of the incorrect use of "this.nonStandard" if "that.nonStandard" is false. Since any other object like window or a PrepOrderCaller instance probably does not have a "nonStandard" field (which would produce a false result for this.nonStandard), likely this code would still have functioned as expected (as long as the previous check clause remains as-is). It would still be good to fix it though to prevent future issues in case other things change.